Split EvohomeRadio from EvohomeSerial
Make general radio message generation and parsing separate from
serial device handling in preparation for adding EvohomeTCP.

Do_Work() still need a lot of work and will hopefully be rationalised
and moved back up to EvohomeRadio once EvohomeTCP is developed. A
bunch of protected stuff in EvohomeRadio can then become private
again.
