+++ CMakeLists.txt
@@ -316,6 +316,7 @@ hardware/EnOceanESP2.cpp
 hardware/EnOceanESP3.cpp
 hardware/Ec3kMeterTCP.cpp
 hardware/EvohomeBase.cpp
+hardware/EvohomeRadio.cpp
 hardware/EvohomeScript.cpp
 hardware/EvohomeSerial.cpp
 hardware/EvohomeWeb.cpp
