+++ main/mainworker.cpp
@@ -13100,8 +13100,3 @@ bool MainWorker::UpdateDevice(const int HardwareID, const std::string &DeviceID,
 	}
 	return true;
 }
-
-
-
-
-
