+++ msbuild/domoticz.vcxproj
@@ -435,6 +435,7 @@
     <ClInclude Include="..\hardware\EnOceanESP3.h" />
     <ClInclude Include="..\hardware\ETH8020.h" />
     <ClInclude Include="..\hardware\EvohomeBase.h" />
+    <ClInclude Include="..\hardware\EvohomeRadio.h" />
     <ClInclude Include="..\hardware\EvohomeScript.h" />
     <ClInclude Include="..\hardware\EvohomeSerial.h" />
     <ClInclude Include="..\hardware\EvohomeWeb.h" />
@@ -683,6 +684,7 @@
     <ClCompile Include="..\hardware\EnOceanESP3.cpp" />
     <ClCompile Include="..\hardware\ETH8020.cpp" />
     <ClCompile Include="..\hardware\EvohomeBase.cpp" />
+    <ClCompile Include="..\hardware\EvohomeRadio.cpp" />
     <ClCompile Include="..\hardware\EvohomeScript.cpp" />
     <ClCompile Include="..\hardware\EvohomeSerial.cpp" />
     <ClCompile Include="..\hardware\EvohomeWeb.cpp" />
